requirements.txt
darkdetect==0.8.0
pillow==10.2.0
pyserial==3.5
